requirements.txt
torch
torchvision
facenet-pytorch
numpy
Pillow
opencv-python
kaggle
